Making previousTicketIndex not fixed-size
- Changed previousTicketIndex will grow as tickets[] grow. Making this array
dynamic should cost less gas usage.
- Replaced uint8 back to uint256 in a couple of places

+++ contracts/solidity/contracts/stubs/KeepRandomBeaconOperatorGroupSelectionStub.sol
@@ -14,7 +14,7 @@ contract KeepRandomBeaconOperatorGroupSelectionStub is KeepRandomBeaconOperator 
         return groupSelection.seed;
     }
 
-    function setGroupSize(uint8 size) public {
+    function setGroupSize(uint256 size) public {
         groupSize = size;
         groupSelection.groupSize = size;
     }
