+++ contracts/solidity/contracts/stubs/KeepRandomBeaconOperatorTicketsOrderingStub.sol
@@ -40,7 +40,7 @@ contract KeepRandomBeaconOperatorTicketsOrderingStub is KeepRandomBeaconOperator
         return groupSelection.previousTicketIndex[higherTicketValueIndex];
     }
 
-    function setGroupSize(uint8 size) public {
+    function setGroupSize(uint256 size) public {
         groupSize = size;
         groupSelection.groupSize = size;
     }
